Fix a major issue with file read. Note #129
This previously set self.file to a pyrax storage object which did not
implement the readline method. As a result Django would fall into
an infinite loop when trying to read the file. This is not a
perfect fix, and it appears to have broken gzip read, but the non-gzip files
are working in django as expected.